fix(assetlib): give the ternary two uint16_t arms, as MSVC requires
A bare 0 promotes the conditional to int, and C4365 is an error there.

+++ libs/assetlib/src/vat_bake.cpp
@@ -63,9 +63,9 @@ namespace assetlib
 					position[axis] =
 						span > 0.0f ?
 							glm::packUnorm1x16((vertex.position[axis] - boundsMin[axis]) / span) :
-							0;
+							uint16_t(0);
 				}
-				position[3] = 65535;
+				position[3] = uint16_t(65535);
 				position += 4;
 
 				// Blending shortens a normal, so it is re-unit here; one a submesh never carried
